# openapi.yaml
openapi: 3.1.0
info:
  title: Canopy Events API
  version: 1.0.0
  summary: Events among friends, for the web page and the iOS and Android apps.
  description: |
    The JSON API behind `events.canopysf.com`. The web pages use it with the
    `canopy_session` cookie; apps use it with `Authorization: Bearer <token>`,
    where the token is a `canopy_session` value. `docs/api.md` in the repo is
    the narrative guide: auth, the RSVP state machine, the visibility rules,
    pagination and errors, with worked examples.

    **Conventions.** JSON in and out. Times are ISO 8601 UTC strings
    (`2026-10-31T03:00:00.000Z`); input times need `Z` or an offset. Ids are
    strings. Errors are `{"error": "<a sentence>", "reason": "<snake_case>"}`
    with the right status: branch on `reason`, show `error` if you like.
    Lists take `?cursor=&limit=` and answer with `nextCursor` (null at the
    end). Every answer is `Cache-Control: no-store`.

    **Other people** are only ever `Person`: id, first and last name, short
    name and photo. Nobody's email, phone, Instagram, Venmo or Cash App is
    ever in an answer, except your own, from `GET /api/v1/me`.
  license:
    name: Private
    identifier: LicenseRef-Proprietary
servers:
  - url: https://events.canopysf.com
    description: Production
  - url: http://localhost:3001
    description: Local development
...
security:
  - bearerAuth: []
  - cookieAuth: []
paths:
...
  /api/v1/events/{id}/wall:
    parameters:
      - $ref: '#/components/parameters/EventId'
    get:
      tags: [Wall]
      operationId: listWall
      summary: The activity wall
      description: |
        Newest first. Readable by anyone who can see the guest list's
        names (`viewer.canSeeGuestList`), since it's full of them; anyone
        else signed in gets `wallVisible: false` and no entries. Entries
        are posts (`type: post`, with `text`) and the server's own, which
        are typed rather than written out, for the app to word: see
        `WallEntry`. `canPost` says whether you may post.
      parameters:
        - $ref: '#/components/parameters/Cursor'
        - $ref: '#/components/parameters/Limit'
      responses:
        '200':
          description: A page of the wall.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Wall' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/SignInRequired' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/EventNotFound' }
        '503': { $ref: '#/components/responses/Unavailable' }
    post:
      tags: [Wall]
      operationId: postToWall
      summary: Post on the wall
      description: |
        Hosts, and anyone whose answer is `going`, `maybe` or
        `waitlisted` (403 `answer_first` otherwise). Plain text, 1 to
        1,000 characters after trimming (400 `bad_text`); it's never
        HTML, so escape it when you show it. Cancelled and finished events
        still take posts. 5 a minute and 100 a day per person (429
        `rate_limited`).
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [text]
              properties:
                text: { type: string, minLength: 1, maxLength: 1000 }
            examples:
              hello:
                value: { text: "Can't wait! I'll bring the cake." }
      responses:
        '201':
          description: The new post.
          content:
            application/json:
              schema:
                type: object
                additionalProperties: false
                required: [entry]
                properties:
                  entry: { $ref: '#/components/schemas/WallEntry' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/SignInRequired' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/EventNotFound' }
        '413': { $ref: '#/components/responses/TooLarge' }
        '429': { $ref: '#/components/responses/RateLimited' }
        '503': { $ref: '#/components/responses/Unavailable' }
  /api/v1/events/{id}/wall/{entryId}:
    parameters:
      - $ref: '#/components/parameters/EventId'
      - name: entryId
        in: path
        required: true
        description: A wall entry's `id`.
        schema: { type: string, pattern: '^[0-9]+$' }
    delete:
      tags: [Wall]
      operationId: deleteWallEntry
      summary: Delete a post or entry
      description: |
        Your own post, or anything on the wall if you're a host (403
        `not_yours` otherwise; `canDelete` on each entry says in
        advance). 404 `entry_not_found` for one that isn't there (already
        deleted, say).
      responses:
        '200':
          description: Deleted.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Ok' }
        '401': { $ref: '#/components/responses/SignInRequired' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '503': { $ref: '#/components/responses/Unavailable' }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: |
        For apps. The token is a `canopy_session` value (43 characters,
        base64url), from the account service's native sign-in (still to be
        built). A request with a bearer header is signed in by it alone,
        never by a cookie, skips the Origin check, and never gets a
        `Set-Cookie` back. A malformed token means signed out.
    cookieAuth:
      type: apiKey
      in: cookie
      name: canopy_session
      description: |
        For the web pages. Set by `account.canopysf.com` for all of
        `canopysf.com`. Changes (POST, PUT, PATCH, DELETE) made with it
        must carry an `Origin` header for a Canopy page, or get 403
        `bad_origin`.
  parameters:
    EventId:
      name: id
      in: path
      required: true
      description: The event's id, 12 characters of base62 (the last part of its link).
      schema: { type: string, pattern: '^[0-9A-Za-z]{12}$' }
      example: 4fQ9xKpL2mZa
    Cursor:
      name: cursor
      in: query
      description: '`nextCursor` from the page before. Leave out (or empty) for the first page.'
      schema: { type: string }
    Limit:
      name: limit
      in: query
      description: How many to a page, 1 to 100.
      schema: { type: integer, minimum: 1, maximum: 100, default: 20 }
...
  responses:
    BadRequest:
      description: |
        Something in the request isn't right: `bad_json`, `bad_title`,
        `bad_starts_at`, `bad_ends_at`, `ends_before_start`,
        `bad_time_zone`, `bad_guest_list_visibility`, `bad_description`,
        `bad_location_name`, `bad_location_address`, `bad_status`,
        `bad_guests`, `too_many_guests`, `bad_guests_allowed`,
        `bad_person_ids`, `bad_person_id`, `bad_text`, `bad_capacity`,
        `bad_image`, `bad_ids`, `bad_platform`, `bad_token`, `bad_cursor`,
        `bad_limit`.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
          example: { error: 'an event has to end after it starts', reason: ends_before_start }
    SignInRequired:
      description: Not signed in (`sign_in_required`), with where to send someone.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
          example:
            error: sign in first
            reason: sign_in_required
            signIn: https://account.canopysf.com/?return=https%3A%2F%2Fevents.canopysf.com%2Fe%2F4fQ9xKpL2mZa
            quickSignUp: https://account.canopysf.com/?quick=1&return=https%3A%2F%2Fevents.canopysf.com%2Fe%2F4fQ9xKpL2mZa
    Forbidden:
      description: |
        Not allowed: `hosts_only`, `creator_only` (co-hosts can't do this),
        `answer_first` (posting before saying you're going or maybe),
        `not_yours` (deleting someone else's post),
        `email_unverified` (with a `verify` link when it's about your own
        email; without one when it's about someone you tried to make a
        co-host), or `bad_origin` (a change made with the cookie from a
        page that isn't Canopy's).
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
          example:
            error: verify your email first
            reason: email_unverified
            verify: https://account.canopysf.com/profile?verify=1&return=https%3A%2F%2Fevents.canopysf.com%2F
    EventNotFound:
      description: No event at that id (`event_not_found`).
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
          example: { error: "there's no event at that link", reason: event_not_found }
    NotFound:
      description: '`event_not_found`, `not_invited`, `person_not_found`, `not_cohost` or `entry_not_found`.'
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
...
    TooLarge:
      description: The body is over 100 KB, or an image over 15 MB (`too_large`).
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    RateLimited:
      description: Over a limit (`rate_limited`). Try again later.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    Unavailable:
      description: Canopy accounts couldn't be reached (`accounts_unreachable`). Try again in a minute.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
  schemas:
    Error:
      type: object
      additionalProperties: false
      required: [error, reason]
      properties:
        error: { type: string, description: A sentence a person can read. }
        reason: { type: string, description: What happened, as a snake_case code to branch on. }
        signIn: { type: string, format: uri, description: 'With `sign_in_required`: where to sign in and come back.' }
        quickSignUp: { type: string, format: uri, description: 'With `sign_in_required`: the quick sign-up (name, email, passkey), for someone new to Canopy.' }
        verify: { type: string, format: uri, description: 'With `email_unverified`: where to prove the email and come back.' }
    Ok:
      type: object
      additionalProperties: false
      required: [ok]
      properties:
        ok: { const: true }
    Person:
      description: |
        Anyone other than you, and the only shape anyone else ever has. A
        deleted account (a former member) is `firstName: "Former member"`,
        `lastName: ""`, `shortName: "Former member"`, `photoUrl: null`.
      type: object
      additionalProperties: false
      required: [id, firstName, lastName, shortName, photoUrl]
      properties:
        id: { type: string }
        firstName: { type: string }
        lastName: { type: string }
        shortName: { type: string, description: First name and last initial, e.g. "Ana L". }
        photoUrl:
          type: [string, 'null']
          format: uri
          description: |
            On account.canopysf.com. Served only with a Canopy session (the
            browser's cookie); null for no photo.
      example:
        id: 6f1c2b9e-4d0a-4a53-9a51-2f7e0c1d8b44
        firstName: Ana
        lastName: Lima
        shortName: Ana L
        photoUrl: https://account.canopysf.com/photo/6f1c2b9e-4d0a-4a53-9a51-2f7e0c1d8b44?v=1759870000000
...
    NextCursor:
      type: [string, 'null']
      description: Pass back as `?cursor=` for the next page; null at the end.
...
    WallEntry:
      description: |
        One thing on the wall. `type` says what, and the app words it:

        | type | means | `person` | `details` |
        |---|---|---|---|
        | `post` | a post, in `text` | who wrote it | null |
        | `going` | they said they're going | who | null |
        | `off_waitlist` | they got a spot from the waitlist | who | null |
        | `time_changed` | a host moved it | the host | the new `startsAt`, `endsAt`, `timeZone` |
        | `place_changed` | a host changed where | the host | the new `locationName`, `locationAddress` |
        | `cancelled` | a host cancelled it | the host | null |
        | `uncancelled` | a host took that back | the host | null |
        | `cohost_added` | they were made a co-host | the new co-host | null |

        Someone has at most one `going` or `off_waitlist` entry, and it's
        gone once they aren't going. Expect new types later: show nothing
        for one you don't know.
      type: object
      additionalProperties: false
      required: [id, type, createdAt, person, text, details, canDelete]
      properties:
        id: { type: string }
        type:
          type: string
          enum: [post, going, off_waitlist, time_changed, place_changed, cancelled, uncancelled, cohost_added]
        createdAt: { type: string, format: date-time }
        person:
          oneOf:
            - $ref: '#/components/schemas/Person'
            - type: 'null'
        text: { type: [string, 'null'], description: A post's text; null for every other type. }
        details:
          oneOf:
            - type: 'null'
            - type: object
              additionalProperties: false
              required: [startsAt, endsAt, timeZone]
              properties:
                startsAt: { type: string, format: date-time }
                endsAt: { type: [string, 'null'], format: date-time }
                timeZone: { type: string }
            - type: object
              additionalProperties: false
              required: [locationName, locationAddress]
              properties:
                locationName: { type: [string, 'null'] }
                locationAddress: { type: [string, 'null'] }
        canDelete: { type: boolean, description: Your own post, or you're a host. }
      example:
        id: '42'
        type: post
        createdAt: '2026-10-08T17:12:00.000Z'
        person: { id: 0b7e5a1f-9c2d-4e8b-8f3a-1d2c3b4a5e6f, firstName: Ben, lastName: Okafor, shortName: Ben O, photoUrl: null }
        text: "Can't wait! I'll bring the cake."
        details: null
        canDelete: false
    Wall:
      type: object
      additionalProperties: false
      required: [wallVisible, entries, canPost, nextCursor]
      properties:
        wallVisible:
          type: boolean
          description: False when you can't see the guest list's names yet; then `entries` is empty.
        entries:
          type: array
          items: { $ref: '#/components/schemas/WallEntry' }
        canPost: { type: boolean }
        nextCursor: { $ref: '#/components/schemas/NextCursor' }
...